Implement sys/seal-status and sys/leader as normal API calls (so that they can be used in namespaces.)
